Commit to the bit
