+++ pyproject.toml
@@ -202,6 +202,7 @@ omit = [
     "docs/*",
     "src/bioregistry/cli.py",
     "src/bioregistry/__main__.py",
+    "src/bioregistry/curation/*.py",
 ]
 
 [tool.coverage.paths]
