+++ src/bioregistry/app/api.py
@@ -3,24 +3,29 @@
 from __future__ import annotations
 
 from collections import defaultdict
-from collections.abc import Mapping
-from typing import Annotated, Any
+from collections.abc import Callable, Mapping
+from typing import Annotated, TypeVar
 
-import yaml
 from curies import Reference
 from curies.mapping_service.utils import handle_header
 from fastapi import APIRouter, Body, Depends, Header, HTTPException, Path, Query, Request
 from fastapi.responses import JSONResponse, Response
 from pydantic import BaseModel, Field
 
-from .utils import FORMAT_MAP, _autocomplete, _search
+from .constants import KEY_TO_MIMETYPE, MIMETYPE_SYNONYM_TO_CANONICAL, MIMETYPE_TO_RDFLIB_FORMAT
+from .responses import TurtleResponse, YAMLResponse
+from .utils import (
+    _autocomplete,
+    _search,
+    get_provider_graph,
+)
 from ..export.rdf_export import (
     collection_to_rdf_str,
     metaresource_to_rdf_str,
     resource_to_rdf_str,
 )
 from ..resource_manager import Manager
-from ..schema import Attributable, Collection, Context, Registry, Resource, sanitize_mapping
+from ..schema import Attributable, Collection, Context, Registry, Resource
 from ..schema.struct import OlsConfig
 from ..schema_utils import (
     read_collections_contributions,
@@ -33,8 +38,6 @@ from ..schema_utils import (
 __all__ = [
     "api_router",
 ]
-
-from ..utils import registry_yaml_dumper
 
 api_router = APIRouter(prefix="/api")
 
@@ -57,82 +60,39 @@ class UnhandledFormat(HTTPException):
         super().__init__(400, f"Bad Accept header: {fmt}")
 
 
-class YAMLResponse(Response):
-    """A custom response encoded in YAML."""
-
-    media_type = "application/yaml"
-
-    def render(self, content: BaseModel | Mapping[str, BaseModel]) -> bytes:
-        """Render content as YAML."""
-        data: dict[str, Any]
-        if isinstance(content, BaseModel):
-            data = content.model_dump(
-                exclude_none=True,
-                exclude_unset=True,
-            )
-        elif isinstance(content, dict):
-            data = sanitize_mapping(content)
-        else:
-            raise TypeError
-        return yaml.safe_dump(
-            data,
-            allow_unicode=True,
-            indent=2,
-        ).encode("utf-8")
-
-
-registry_yaml_dumper()
-
-ACCEPT_HEADER = Header(default=None)
-FORMAT_QUERY = Query(
-    title="Format", default=None, description=f"The return format, one of: {list(FORMAT_MAP)}"
-)
-#: A mapping of mimetypes to RDFLib formats
-RDF_MEDIA_TYPES = {
-    "text/turtle": "turtle",
-    "application/ld+json": "json-ld",
-    "application/rdf+xml": "xml",
-    "text/n3": "n3",
-}
-CONTENT_TYPE_SYNONYMS = {
-    "text/json": "application/json",
-    "text/yaml": "application/yaml",
-}
+Accept = Annotated[str | None, Header()]
+Format = Annotated[
+    str | None,
+    Query(
+        title="Format",
+        description=f"The return format, one of: {list(KEY_TO_MIMETYPE)}",
+    ),
+]
 
 
 def _handle_formats(accept: str | None, fmt: str | None) -> str:
     if fmt:
-        if fmt not in FORMAT_MAP:
+        if fmt not in KEY_TO_MIMETYPE:
             raise HTTPException(
-                400, f"bad query parameter format={fmt}. Should be one of {list(FORMAT_MAP)}"
+                400, f"bad query parameter format={fmt}. Should be one of {list(KEY_TO_MIMETYPE)}"
             )
-        return FORMAT_MAP[fmt]
+        return KEY_TO_MIMETYPE[fmt]
     if not accept:
         return "application/json"
     for header in handle_header(accept):
-        if header in CONTENT_TYPE_SYNONYMS:
-            return CONTENT_TYPE_SYNONYMS[header]
-        if header in RDF_MEDIA_TYPES or header in CONTENT_TYPE_SYNONYMS.values():
+        if header in MIMETYPE_SYNONYM_TO_CANONICAL:
+            return MIMETYPE_SYNONYM_TO_CANONICAL[header]
+        if header in MIMETYPE_TO_RDFLIB_FORMAT or header in MIMETYPE_SYNONYM_TO_CANONICAL.values():
             return header
     return "application/json"
 
 
 @api_router.get("/registry", response_model=Mapping[str, Resource], tags=["resource"])
 def get_resources(
-    manager: DependsManager,
-    accept: str | None = ACCEPT_HEADER,
-    format: str | None = FORMAT_QUERY,
+    manager: DependsManager, accept: Accept = None, format: Format = None
 ) -> Response | dict[str, Resource]:
     """Get all resources."""
-    accept = _handle_formats(accept, format)
-    if accept == "application/json":
-        return manager.registry
-    elif accept == "application/yaml":
-        return YAMLResponse(manager.registry)
-    elif accept in RDF_MEDIA_TYPES:
-        raise NotImplementedError
-    else:
-        raise UnhandledFormat(accept)
+    return serialize_model_fastapi(manager, accept, format, manager.registry)
 
 
 @api_router.get(
@@ -143,7 +103,7 @@ def get_resources(
         200: {
             "content": {
                 "application/yaml": {},
-                **{k: {} for k in RDF_MEDIA_TYPES},
+                **{k: {} for k in MIMETYPE_TO_RDFLIB_FORMAT},
             },
         },
     },
@@ -155,26 +115,15 @@ def get_resource(
     prefix: str = Path(
         title="Prefix", description="The internal prefix for the entry", examples=["doid"]
     ),
-    accept: str = ACCEPT_HEADER,
-    format: str = FORMAT_QUERY,
+    accept: Accept = None,
+    format: Format = None,
 ) -> Response | Resource:
     """Get a resource."""
     resource = manager.get_resource(prefix)
     if resource is None:
         raise HTTPException(status_code=404, detail=f"Prefix not found: {prefix}")
     resource = manager.rasterized_resource(resource)
-    accept = _handle_formats(accept, format)
-    if accept == "application/json":
-        return resource
-    elif accept == "application/yaml":
-        return YAMLResponse(resource)
-    elif accept in RDF_MEDIA_TYPES:
-        return Response(
-            resource_to_rdf_str(resource, fmt=RDF_MEDIA_TYPES[accept], manager=manager),
-            media_type=accept,
-        )
-    else:
-        raise UnhandledFormat(format)
+    return serialize_model_fastapi(manager, accept, format, resource, func=resource_to_rdf_str)
 
 
 @api_router.get(
@@ -209,26 +158,21 @@ def get_resource_ols_config(
 )
 def get_metaresources(
     manager: DependsManager,
-    accept: str | None = ACCEPT_HEADER,
-    format: str | None = FORMAT_QUERY,
+    accept: Accept = None,
+    format: Format = None,
 ) -> Response | Mapping[str, Registry]:
     """Get all registries."""
-    accept = _handle_formats(accept, format)
-    if accept == "application/json":
-        return manager.metaregistry
-    elif accept == "application/yaml":
-        return YAMLResponse(manager.metaregistry)
-    elif accept in RDF_MEDIA_TYPES:
-        raise NotImplementedError
-    else:
-        raise UnhandledFormat(accept)
+    return serialize_model_fastapi(manager, accept, format, manager.metaregistry)
 
 
-METAPREFIX_PATH = Path(
-    title="Metaprefix",
-    description="The metaprefix for the external registry",
-    examples=["n2t"],
-)
+Metaprefix = Annotated[
+    str,
+    Path(
+        title="Metaprefix",
+        description="The metaprefix for the external registry",
+        examples=["n2t"],
+    ),
+]
 
 
 @api_router.get(
@@ -240,7 +184,7 @@ METAPREFIX_PATH = Path(
         200: {
             "content": {
                 "application/yaml": {},
-                **{k: {} for k in RDF_MEDIA_TYPES},
+                **{k: {} for k in MIMETYPE_TO_RDFLIB_FORMAT},
             },
         },
     },
@@ -249,30 +193,44 @@ METAPREFIX_PATH = Path(
 )
 def get_metaresource(
     manager: DependsManager,
-    metaprefix: str = METAPREFIX_PATH,
-    accept: str = ACCEPT_HEADER,
-    format: str = FORMAT_QUERY,
+    metaprefix: Metaprefix,
+    accept: Accept = None,
+    format: Format = None,
 ) -> Response | Registry:
     """Get all registries."""
     metaresource = manager.get_registry(metaprefix)
     if metaresource is None:
         raise HTTPException(status_code=404, detail=f"Registry not found: {metaprefix}")
+    return serialize_model_fastapi(
+        manager, accept, format, metaresource, func=metaresource_to_rdf_str
+    )
+
+
+X = TypeVar("X", bound=BaseModel | Mapping[str, BaseModel])
+
+
+def serialize_model_fastapi(
+    manager: Manager,
+    accept: str | None,
+    format: str | None,
+    model: X,
+    func: Callable[[X, Manager, str | None], str] | None = None,
+) -> X | Response:
+    """Serialize a model in FastAPI."""
     accept = _handle_formats(accept, format)
     if accept == "application/json":
-        return metaresource
+        return model
     elif accept == "application/yaml":
-        return YAMLResponse(metaresource)
-    elif accept in RDF_MEDIA_TYPES:
+        return YAMLResponse(model)
+    elif accept in MIMETYPE_TO_RDFLIB_FORMAT:
+        if not func:
+            raise NotImplementedError
         return Response(
-            metaresource_to_rdf_str(
-                metaresource,
-                fmt=RDF_MEDIA_TYPES[accept],
-                manager=manager,
-            ),
+            func(model, manager, MIMETYPE_TO_RDFLIB_FORMAT[accept]),
             media_type=accept,
         )
     else:
-        raise UnhandledFormat(format)
+        raise HTTPException(400, f"Bad Accept header: {accept}")
 
 
 @api_router.get(
@@ -281,8 +239,7 @@ def get_metaresource(
     tags=["metaresource"],
 )
 def get_external_registry_slim(
-    manager: DependsManager,
-    metaprefix: str = METAPREFIX_PATH,
+    manager: DependsManager, metaprefix: Metaprefix
 ) -> dict[str, Resource]:
     """Get a slim version of the registry with only resources mapped to the given external registry."""
     return {
@@ -319,8 +276,8 @@ class MappingResponse(BaseModel):
 )
 def get_metaresource_external_mappings(
     manager: DependsManager,
-    metaprefix: str = METAPREFIX_PATH,
-    target: str = Path(title="target metaprefix"),
+    metaprefix: Metaprefix,
+    target: Annotated[str, Path(title="target metaprefix")],
 ) -> MappingResponse | tuple[dict[str, str], int]:
     """Get mappings between two external prefixes."""
     try:
@@ -346,9 +303,7 @@ def get_metaresource_external_mappings(
     response_model=Mapping[str, str],
     tags=["metaresource"],
 )
-def get_metaresource_mappings(
-    manager: DependsManager, metaprefix: str = METAPREFIX_PATH
-) -> dict[str, str]:
+def get_metaresource_mappings(manager: DependsManager, metaprefix: Metaprefix) -> dict[str, str]:
     """Get mappings from internal to external prefixes for a given external registry."""
     if metaprefix not in manager.metaregistry:
         raise HTTPException(404, detail=f"Invalid metaprefix: {metaprefix}")
@@ -357,20 +312,10 @@ def get_metaresource_mappings(
 
 @api_router.get("/collection", response_model=Mapping[str, Collection], tags=["collection"])
 def get_collections(
-    manager: DependsManager,
-    accept: str | None = ACCEPT_HEADER,
-    format: str | None = FORMAT_QUERY,
+    manager: DependsManager, accept: Accept = None, format: Format = None
 ) -> Response | dict[str, Collection]:
     """Get all collections."""
-    accept = _handle_formats(accept, format)
-    if accept == "application/json":
-        return manager.collections
-    elif accept == "application/yaml":
-        return YAMLResponse(manager.collections)
-    elif accept in RDF_MEDIA_TYPES:
-        raise NotImplementedError
-    else:
-        raise HTTPException(400, f"Bad Accept header: {accept}")
+    return serialize_model_fastapi(manager, accept, format, manager.collections)
 
 
 COLLECTION_IDENTIFIER = Path(
@@ -388,7 +333,7 @@ COLLECTION_IDENTIFIER = Path(
         200: {
             "content": {
                 "application/yaml": {},
-                **{k: {} for k in RDF_MEDIA_TYPES},
+                **{k: {} for k in MIMETYPE_TO_RDFLIB_FORMAT},
             },
         },
     },
@@ -398,8 +343,8 @@ COLLECTION_IDENTIFIER = Path(
 def get_collection(
     manager: DependsManager,
     identifier: Annotated[str, COLLECTION_IDENTIFIER],
-    accept: str | None = ACCEPT_HEADER,
-    format: str | None = FORMAT_QUERY,
+    accept: Accept = None,
+    format: Format = None,
 ) -> Response | Collection:
     """Get a collection."""
     collection = manager.collections.get(identifier)
@@ -407,22 +352,7 @@ def get_collection(
         raise HTTPException(status_code=404, detail=f"Collection not found: {identifier}")
     if accept == "x-bioregistry-context" or format == "context":
         return JSONResponse(collection.as_context_jsonld())
-    accept = _handle_formats(accept, format)
-    if accept == "application/json":
-        return collection
-    elif accept == "application/yaml":
-        return YAMLResponse(collection)
-    elif accept in RDF_MEDIA_TYPES:
-        return Response(
-            collection_to_rdf_str(
-                collection,
-                fmt=RDF_MEDIA_TYPES[accept],
-                manager=manager,
-            ),
-            media_type=accept,
-        )
-    else:
-        raise HTTPException(400, f"Bad Accept header: {accept}")
+    return serialize_model_fastapi(manager, accept, format, collection, func=collection_to_rdf_str)
 
 
 class CollectionMappingResult(BaseModel):
@@ -491,7 +421,7 @@ class OLSConfigurations(BaseModel):
 )
 def get_ols_configurations(
     manager: DependsManager,
-    prefixes: Annotated[str, Path(..., examples=["cl,doid,mondo"])],
+    prefixes: Annotated[str, Path(examples=["cl,doid,mondo"])],
 ) -> OLSConfigurations:
     """Get OLS configurations for multiple prefixes given as a query parameter."""
     return _get_multiple_ols_configurations(prefixes.split(","), manager)
@@ -529,18 +459,10 @@ def _get_multiple_ols_configurations(prefixes: list[str], manager: Manager) -> O
 
 @api_router.get("/context", response_model=Mapping[str, Context], tags=["context"])
 def get_contexts(
-    manager: DependsManager,
-    accept: str | None = ACCEPT_HEADER,
-    format: str | None = FORMAT_QUERY,
+    manager: DependsManager, accept: Accept = None, format: Format = None
 ) -> Response | dict[str, Context]:
     """Get all context."""
-    accept = _handle_formats(accept, format)
-    if accept == "application/json":
-        return manager.contexts
-    elif accept == "application/yaml":
-        return YAMLResponse(manager.contexts)
-    else:
-        raise HTTPException(400, f"Bad Accept header: {accept}")
+    return serialize_model_fastapi(manager, accept, format, manager.contexts)
 
 
 @api_router.get("/context/{identifier}", response_model=Context, tags=["context"])
@@ -557,41 +479,33 @@ def get_context(
 
 @api_router.get("/contributors", response_model=Mapping[str, Attributable], tags=["contributor"])
 def get_contributors(
-    manager: DependsManager,
-    accept: str | None = ACCEPT_HEADER,
-    format: str | None = FORMAT_QUERY,
+    manager: DependsManager, accept: Accept = None, format: Format = None
 ) -> Response | Mapping[str, Attributable]:
     """Get all context."""
     contributors = manager.read_contributors()
-    accept = _handle_formats(accept, format)
-    if accept == "application/json":
-        return contributors
-    elif accept == "application/yaml":
-        return YAMLResponse(contributors)
-    else:
-        raise HTTPException(400, f"Bad Accept header: {accept}")
+    return serialize_model_fastapi(manager, accept, format, contributors)
 
 
 class ContributorResponse(BaseModel):
     """A response with information about a contributor."""
 
     contributor: Attributable
-    prefix_contributions: set[str]
-    prefix_reviews: set[str]
-    prefix_contacts: set[str]
-    registries: set[str]
-    collections: set[str]
+    prefix_contributions: list[str]
+    prefix_reviews: list[str]
+    prefix_contacts: list[str]
+    registries: list[str]
+    collections: list[str]
 
 
 @api_router.get("/contributor/{orcid}", response_model=ContributorResponse, tags=["contributor"])
 def get_contributor(
     manager: DependsManager,
-    orcid: Annotated[str, Path(..., title="Open Researcher and Contributor Identifier")],
+    orcid: Annotated[str, Path(title="Open Researcher and Contributor Identifier")],
 ) -> ContributorResponse:
     """Get all context."""
     author = manager.read_contributors().get(orcid)
     if author is None:
-        raise HTTPException(404, f"No contributor with orcid: {orcid}")
+        raise HTTPException(404, f"No contributor with ORCiD: {orcid}")
     return ContributorResponse(
         contributor=author,
         prefix_contributions=sorted(read_prefix_contributions(manager.registry).get(orcid, [])),
@@ -612,7 +526,12 @@ class IdentifierResponse(BaseModel):
 @api_router.get(
     "/reference/{prefix}:{identifier:path}", response_model=IdentifierResponse, tags=["reference"]
 )
-def get_reference(manager: DependsManager, prefix: str, identifier: str) -> IdentifierResponse:
+def get_reference(
+    manager: DependsManager,
+    prefix: str,
+    identifier: str,
+    format: Format = None,
+) -> IdentifierResponse | YAMLResponse | TurtleResponse:
     """Look up information on the reference."""
     # see https://fastapi.tiangolo.com/tutorial/path-params/#path-parameters-containing-paths
     # for more understanding on how the identifier:path handling works
@@ -630,39 +549,52 @@ def get_reference(manager: DependsManager, prefix: str, identifier: str) -> Iden
     if not providers:
         raise HTTPException(404, f"no providers available for {resource.get_curie(identifier)}")
 
-    return IdentifierResponse(
-        query=Reference(prefix=prefix, identifier=identifier),
-        providers=providers,
-    )
+    reference = Reference(prefix=prefix, identifier=identifier)
+
+    if format == "json" or format is None:
+        return IdentifierResponse(query=reference, providers=providers)
+    elif format == "yaml":
+        x = IdentifierResponse(query=reference, providers=providers)
+        return YAMLResponse(x)
+    elif format == "turtle":
+        return TurtleResponse(
+            get_provider_graph(manager, reference.prefix, reference.identifier, providers)
+        )
+    else:
+        raise HTTPException(404, f"invalid format: {format}")
 
 
 class URIResponse(BaseModel):
     """A response for looking up a reference."""
 
-    uri: str = Field(
-        ..., description="The query URI", examples=["http://id.nlm.nih.gov/mesh/C063233"]
-    )
-    reference: Reference = Field(
-        ...,
-        description="The compact URI (CURIE)",
-        examples=[Reference(prefix="mesh", identifier="C063233")],
-    )
-    providers: Mapping[str, str] = Field(
-        ...,
-        description="Equivalent URIs",
-        examples=[
-            {
-                "default": "https://meshb.nlm.nih.gov/record/ui?ui=C063233",
-                "rdf": "http://id.nlm.nih.gov/mesh/C063233",
-            }
-        ],
-    )
+    uri: Annotated[
+        str, Field(description="The query URI", examples=["http://id.nlm.nih.gov/mesh/C063233"])
+    ]
+    reference: Annotated[
+        Reference,
+        Field(
+            description="The compact URI (CURIE)",
+            examples=[Reference(prefix="mesh", identifier="C063233")],
+        ),
+    ]
+    providers: Annotated[
+        Mapping[str, str],
+        Field(
+            description="Equivalent URIs",
+            examples=[
+                {
+                    "default": "https://meshb.nlm.nih.gov/record/ui?ui=C063233",
+                    "rdf": "http://id.nlm.nih.gov/mesh/C063233",
+                }
+            ],
+        ),
+    ]
 
 
 class URIQuery(BaseModel):
     """A query for parsing a URI."""
 
-    uri: str = Field(..., examples=["http://id.nlm.nih.gov/mesh/C063233"])
+    uri: Annotated[str, Field(examples=["http://id.nlm.nih.gov/mesh/C063233"])]
 
 
 @api_router.post(
@@ -670,9 +602,7 @@ class URIQuery(BaseModel):
 )
 def post_parse_uri(
     manager: DependsManager,
-    query: Annotated[
-        URIQuery, Body(..., examples=[URIQuery(uri="http://id.nlm.nih.gov/mesh/C063233")])
-    ],
+    query: Annotated[URIQuery, Body(examples=[URIQuery(uri="http://id.nlm.nih.gov/mesh/C063233")])],
 ) -> URIResponse:
     """Parse a URI, return a CURIE, and all equivalent URIs."""
     prefix, identifier = manager.parse_uri(query.uri)
@@ -690,7 +620,7 @@ def post_parse_uri(
 def generate_context_json_ld(
     manager: DependsManager,
     prefix: Annotated[
-        list[str], Query(..., description="The prefix for the entry. Can be given multiple.")
+        list[str], Query(description="The prefix for the entry. Can be given multiple.")
     ],
 ) -> JSONResponse:
     """Generate an *ad-hoc* context JSON-LD file from the given parameters.
