+++ src/bioregistry/app/constants.py
@@ -1,10 +1,19 @@
 """Large data chunks for the web app."""
 
+from __future__ import annotations
+
+from pathlib import Path
+
 from .. import version
 from ..constants import INTERNAL_LABEL, INTERNAL_REPOSITORY_BLOB
 
 __all__ = [
     "BIOSCHEMAS",
+    "FORMATS",
+    "KEY_TO_MIMETYPE",
+    "MIMETYPE_SYNONYM_TO_CANONICAL",
+    "MIMETYPE_TO_RDFLIB_FORMAT",
+    "TEMPLATES_DIRECTORY",
 ]
 
 BIOSCHEMAS = {
@@ -66,3 +75,33 @@ KEY_B = "0d15"
 KEY_C = "4e75"
 KEY_D = "9b71"
 KEY_E = "b9e7da1d6ac3"
+
+#: A mapping of mimetypes to RDFLib formats
+MIMETYPE_TO_RDFLIB_FORMAT = {
+    "text/turtle": "turtle",
+    "application/ld+json": "json-ld",
+    "application/rdf+xml": "xml",
+    "text/n3": "n3",
+}
+
+#: keys exposed through the API to map back to mimetypes
+KEY_TO_MIMETYPE = {
+    "json": "application/json",
+    "yml": "application/yaml",
+    "yaml": "application/yaml",
+    "turtle": "text/turtle",
+    "jsonld": "application/ld+json",
+    "json-ld": "application/ld+json",
+    "rdf": "application/rdf+xml",
+    "n3": "text/n3",
+}
+
+MIMETYPE_SYNONYM_TO_CANONICAL = {
+    "text/json": "application/json",
+    "text/yaml": "application/yaml",
+}
+TEMPLATES_DIRECTORY = Path(__file__).parent.resolve().joinpath("templates")
+FORMATS = [
+    ("JSON", "json"),
+    ("YAML", "yaml"),
+]
