+++ src/bioregistry/app/ui.py
@@ -2,44 +2,30 @@
 
 from __future__ import annotations
 
-import datetime
 import itertools as itt
 import json
-import platform
-from collections import Counter, defaultdict
+from collections import defaultdict
 from collections.abc import Iterable
 from operator import attrgetter
-from pathlib import Path
 from typing import cast
 
 import flask
 import werkzeug
 from curies import Reference
-from flask import (
-    Blueprint,
-    abort,
-    current_app,
-    jsonify,
-    redirect,
-    render_template,
-    request,
-    url_for,
-)
+from flask import abort, current_app, jsonify, redirect, render_template, request, url_for
 
+from .components.base import ui_blueprint
+from .components.resolver import _clean_reference, resolve
+from .constants import FORMATS
 from .proxies import manager
 from .utils import (
+    ResponseWrapperError,
     _get_resource_providers,
-    _normalize_prefix_or_404,
     get_accept_media_type,
     serialize_model,
 )
-from .. import version
-from ..constants import INTERNAL_LABEL, INTERNAL_METAPREFIX, NDEX_UUID, NFDI_ROR
-from ..export.rdf_export import (
-    collection_to_rdf_str,
-    metaresource_to_rdf_str,
-    resource_to_rdf_str,
-)
+from ..constants import INTERNAL_LABEL, INTERNAL_METAPREFIX, NFDI_ROR
+from ..export.rdf_export import collection_to_rdf_str, metaresource_to_rdf_str
 from ..schema import (
     Context,
     Registry,
@@ -50,10 +36,8 @@ from ..schema import (
     get_json_schema,
     schema_status_map,
 )
-from ..schema.constants import SCHEMA_TERMS
 from ..schema.struct import Collection, Organization, filter_collections
 from ..schema_utils import (
-    get_collection_mappings,
     read_collections_contributions,
     read_context_contributions,
     read_prefix_contacts,
@@ -64,17 +48,7 @@ from ..schema_utils import (
 )
 from ..utils import curie_to_str
 
-__all__ = [
-    "ui_blueprint",
-]
-
-TEMPLATES = Path(__file__).parent.resolve().joinpath("templates")
-ui_blueprint = Blueprint("metaregistry_ui", __name__, template_folder=TEMPLATES.as_posix())
-
-FORMATS = [
-    ("JSON", "json"),
-    ("YAML", "yaml"),
-]
+__all__ = ["ui_blueprint"]
 
 
 @ui_blueprint.route("/registry/")
@@ -116,86 +90,6 @@ def collections() -> str:
         formats=FORMATS,
         ror=ror,
         organization=organization,
-    )
-
-
-@ui_blueprint.route("/registry/<prefix>")
-def resource(prefix: str) -> str | werkzeug.Response | tuple[str, int]:
-    """Serve a resource page."""
-    norm_prefix = _normalize_prefix_or_404(prefix, "." + resource.__name__)
-    if not isinstance(norm_prefix, str):
-        return norm_prefix
-    _resource = manager.get_resource(norm_prefix)
-    if _resource is None:
-        raise RuntimeError
-    accept = get_accept_media_type()
-    if accept != "text/html":
-        return serialize_model(_resource, resource_to_rdf_str, negotiate=True)
-
-    example = _resource.get_example()
-    example_curie = _resource.get_example_curie(use_preferred=True)
-    example_extras = _resource.get_example_extras()
-    example_curie_extras = [
-        _resource.get_curie(example_extra, use_preferred=True) for example_extra in example_extras
-    ]
-    name_pack = manager._repack(_resource.get_name(provenance=True))
-    return render_template(
-        "resource.html",
-        zip=zip,
-        prefix=prefix,
-        resource=_resource,
-        bioschemas=json.dumps(_resource.get_bioschemas_jsonld(), ensure_ascii=False),
-        name_pack=name_pack,
-        example=example,
-        example_extras=example_extras,
-        example_curie=example_curie,
-        example_curie_extras=example_curie_extras,
-        mappings=[
-            {
-                "metaprefix": metaprefix,
-                "metaresource": manager.get_registry(metaprefix),
-                "xref": xref,
-                "homepage": manager.get_registry_homepage(metaprefix),
-                "name": manager.get_registry_name(metaprefix),
-                "short_name": manager.get_registry_short_name(metaprefix),
-                "uri": manager.get_registry_provider_uri_format(metaprefix, xref),
-            }
-            for metaprefix, xref in _resource.get_mappings().items()
-        ],
-        synonyms=_resource.get_synonyms(),
-        homepage=_resource.get_homepage(),
-        repository=_resource.get_repository(),
-        pattern=manager.get_pattern(prefix),
-        curie_pattern=manager.get_curie_pattern(prefix, use_preferred=True),
-        version=_resource.get_version(),
-        has_no_terms=manager.has_no_terms(prefix),
-        obo_download=_resource.get_download_obo(),
-        owl_download=_resource.get_download_owl(),
-        json_download=_resource.get_download_obograph(),
-        rdf_download=_resource.get_download_rdf(),
-        skos_download=_resource.get_download_skos(),
-        jskos_download=_resource.get_download_jskos(),
-        namespace_in_lui=_resource.get_namespace_in_lui(),
-        deprecated=manager.is_deprecated(prefix),
-        contact=_resource.get_contact(),
-        banana=_resource.get_banana(),
-        description=manager.get_description(prefix, use_markdown=True),
-        appears_in=manager.get_appears_in(prefix),
-        depends_on=manager.get_depends_on(prefix),
-        has_canonical=manager.get_has_canonical(prefix),
-        canonical_for=manager.get_canonical_for(prefix),
-        provides=manager.get_provides_for(prefix),
-        provided_by=manager.get_provided_by(prefix),
-        part_of=manager.get_part_of(prefix),
-        has_parts=manager.get_has_parts(prefix),
-        in_collections=manager.get_in_collections(prefix),
-        providers=None if example is None else _get_resource_providers(prefix, example),
-        formats=[
-            *FORMATS,
-            ("RDF (turtle)", "turtle"),
-            ("RDF (JSON-LD)", "jsonld"),
-            ("RDF (n3)", "n3"),
-        ],
     )
 
 
@@ -315,61 +209,6 @@ def context(identifier: str) -> str:
     )
 
 
-class ResponseWrapperError(ValueError):
-    """An exception that helps with code reuse that returns multiple value types."""
-
-    def __init__(self, response: str | werkzeug.Response, code: int | None = None) -> None:
-        """Instantiate this "exception", which is a tricky way of writing a macro."""
-        self.response = response
-        self.code = code
-
-    def get_value(self) -> tuple[str | werkzeug.Response, int] | str | werkzeug.Response:
-        """Get either the response, or a pair of response + code if a code is available."""
-        if self.code is not None:
-            return self.response, self.code
-        return self.response
-
-
-def _clean_reference(prefix: str, identifier: str | None = None) -> tuple[Resource, str]:
-    if ":" in prefix:
-        # A colon might appear in the prefix if there are multiple colons
-        # in the CURIE, since Flask/Werkzeug parses from right to left.
-        # This block reorganizes the parts of the CURIE based on that assumption
-        prefix, middle = prefix.split(":", 1)
-        if identifier:
-            identifier = f"{middle}:{identifier}"
-        else:
-            identifier = middle  # not sure how this could happen, though
-
-    _resource = manager.get_resource(prefix)
-    if _resource is None:
-        raise ResponseWrapperError(
-            render_template(
-                "resolve_errors/missing_prefix.html", prefix=prefix, identifier=identifier
-            ),
-            404,
-        )
-    if identifier is None:
-        raise ResponseWrapperError(
-            redirect(url_for("." + resource.__name__, prefix=_resource.prefix))
-        )
-
-    identifier = _resource.standardize_identifier(identifier)
-    pattern = _resource.get_pattern()
-    if pattern and not _resource.is_valid_identifier(identifier):
-        raise ResponseWrapperError(
-            render_template(
-                "resolve_errors/invalid_identifier.html",
-                prefix=prefix,
-                identifier=identifier,
-                pattern=pattern,
-            ),
-            404,
-        )
-
-    return _resource, identifier
-
-
 @ui_blueprint.route("/reference/<prefix>:<path:identifier>")
 @ui_blueprint.route("/reference/<prefix>:/<path:identifier>")  # ARK hack, see below
 def reference(
@@ -386,61 +225,11 @@ def reference(
         name=_resource.get_name(),
         identifier=identifier,
         providers=_get_resource_providers(_resource.prefix, identifier),
-        formats=FORMATS,
+        formats=[
+            *FORMATS,
+            ("RDF (turtle)", "turtle"),
+        ],
     )
-
-
-#: this is a hack to make it work when the luid starts with a slash for
-#: ARK, since ARK doesn't actually require a slash. Will break
-#: if there are other LUIDs that actualyl require a slash in front
-ark_hacked_route = ui_blueprint.route("/<prefix>:/<path:identifier>")
-
-
-@ui_blueprint.route("/<prefix>")
-@ui_blueprint.route("/<prefix>:<path:identifier>")
-@ark_hacked_route
-def resolve(
-    prefix: str, identifier: str | None = None
-) -> str | werkzeug.Response | tuple[str | werkzeug.Response, int]:
-    """Resolve a CURIE.
-
-    The following things can make a CURIE unable to resolve:
-
-    1. The prefix is not registered
-    2. The prefix has a validation pattern and the identifier does not match it
-    3. There are no providers available for the URL
-    """
-    try:
-        _resource, identifier = _clean_reference(prefix, identifier)
-    except ResponseWrapperError as rw:
-        return rw.get_value()
-    url = manager.get_iri(
-        _resource.prefix,
-        identifier,
-        use_bioregistry_io=False,
-        provider=request.args.get("provider"),
-    )
-    if not url:
-        return (
-            render_template(
-                "resolve_errors/missing_providers.html",
-                prefix=_resource.prefix,
-                identifier=identifier,
-            ),
-            404,
-        )
-    try:
-        # TODO remove any garbage characters?
-        return redirect(url)
-    except ValueError:  # headers could not be constructed
-        return (
-            render_template(
-                "resolve_errors/disallowed_identifier.html",
-                prefix=_resource.prefix,
-                identifier=identifier,
-            ),
-            404,
-        )
 
 
 @ui_blueprint.route("/metaregistry/<metaprefix>/<metaidentifier>")
@@ -466,18 +255,6 @@ def metaresolve(
             f" {list(manager.get_registry_invmap(metaprefix))}",
         )
     return redirect(url_for(f".{resolve.__name__}", prefix=prefix, identifier=identifier))
-
-
-@ui_blueprint.route("/resolve/github/issue/<owner>/<repository>/<int:issue>")
-def github_resolve_issue(owner: str, repository: str, issue: str) -> werkzeug.Response:
-    """Redirect to an issue on GitHub."""
-    return redirect(f"https://github.com/{owner}/{repository}/issues/{issue}")
-
-
-@ui_blueprint.route("/resolve/github/pull/<owner>/<repository>/<int:pull>")
-def github_resolve_pull(owner: str, repository: str, pull: int) -> werkzeug.Response:
-    """Redirect to a pull request on GitHub."""
-    return redirect(f"https://github.com/{owner}/{repository}/pull/{pull}")
 
 
 @ui_blueprint.route("/contributors/")
@@ -563,12 +340,6 @@ def home() -> str:
     )
 
 
-@ui_blueprint.route("/summary")
-def summary() -> str:
-    """Render the summary page."""
-    return render_template("meta/summary.html")
-
-
 @ui_blueprint.route("/related")
 def related() -> str:
     """Render the related page."""
@@ -584,78 +355,10 @@ def related() -> str:
     )
 
 
-@ui_blueprint.route("/download")
-def download() -> str:
-    """Render the download page."""
-    return render_template("meta/download.html", ndex_uuid=NDEX_UUID)
-
-
-@ui_blueprint.route("/acknowledgements")
-def acknowledgements() -> str:
-    """Render the acknowledgements page."""
-    return render_template(
-        "meta/acknowledgements.html",
-        registries=sorted(manager.metaregistry.values(), key=attrgetter("name")),
-    )
-
-
-_VERSION = version.get_version()
-_GIT_HASH = version.get_git_hash()
-_PLATFORM = platform.platform()
-_PLATFORM_VERSION = platform.version()
-_PYTHON_VERSION = platform.python_version()
-_DEPLOYED = datetime.datetime.now()
-
-
-@ui_blueprint.route("/sustainability")
-def sustainability() -> str:
-    """Render the sustainability page."""
-    return render_template(
-        "meta/sustainability.html",
-        software_version=_VERSION,
-        software_git_hash=_GIT_HASH,
-        platform=_PLATFORM,
-        platform_version=_PLATFORM_VERSION,
-        python_version=_PYTHON_VERSION,
-        deployed=_DEPLOYED,
-    )
-
-
-@ui_blueprint.route("/usage")
-def usage() -> str:
-    """Render the programmatic usage page."""
-    resource = manager.get_resource(current_app.config["METAREGISTRY_EXAMPLE_PREFIX"])
-    return render_template("meta/access.html", resource=resource)
-
-
-@ui_blueprint.route("/.well-known/funding-manifest-urls")
-def funding_manifest_urls() -> werkzeug.Response:
-    """Render the FLOSS Fund page, described by https://floss.fund/funding-manifest/."""
-    return current_app.send_static_file("funding-manifest-urls.txt")
-
-
-@ui_blueprint.route("/schema/")
-def schema() -> str:
-    """Render the RDF schema."""
-    return render_template("meta/schema.html", terms=SCHEMA_TERMS)
-
-
 @ui_blueprint.route("/schema.json")
 def json_schema() -> werkzeug.Response:
     """Return the JSON schema."""
     return jsonify(get_json_schema())  # type:ignore
-
-
-@ui_blueprint.route("/highlights/twitter")
-def highlights_twitter() -> flask.Response:
-    """Render a HTTP 410 Gone (forever) for twitter content, which is no longer supported."""
-    return flask.Response("twitter content is no longer tracked", status=410)
-
-
-@ui_blueprint.route("/highlights/relations")
-def highlights_relations() -> str:
-    """Render the relations highlights page."""
-    return render_template("highlights/relations.html")
 
 
 @ui_blueprint.route("/keywords")
@@ -728,78 +431,3 @@ def show_organization(curie: str) -> str:
 def apidocs() -> werkzeug.Response:
     """Render api documentation page."""
     return redirect("/docs")
-
-
-@ui_blueprint.route("/nfdi")
-@ui_blueprint.route("/nfdi/")
-def show_nfdi() -> str:
-    """Render the NFDI dashboard page."""
-    nfdi_collections = {
-        c.identifier: c
-        for c in manager.collections.values()
-        if c.has_organization_with_ror(NFDI_ROR)
-    }
-    tib_collection_mappings = get_collection_mappings("tib.collection")
-    bartoc_collection_mappings = get_collection_mappings("bartoc")
-    collection_to_tib_opportunities = defaultdict(list)
-    collection_to_license_needs_curation = defaultdict(list)
-    collection_to_domain_needs_curation = defaultdict(list)
-    collection_to_download_need_curation = defaultdict(list)
-    tib_opportunities = set()
-    for collection_ in nfdi_collections.values():
-        for prefix in collection_.get_prefixes():
-            if prefix == "bioregistry":
-                continue
-            resource_ = manager.get_resource(prefix, strict=True)
-            if not resource_.get_mapped_prefix("tib") and resource_.has_download():
-                collection_to_tib_opportunities[collection_.identifier].append(prefix)
-                tib_opportunities.add(prefix)
-            if not resource_.get_license():
-                collection_to_license_needs_curation[collection_.identifier].append(resource)
-            if not resource_.domain:
-                collection_to_domain_needs_curation[collection_.identifier].append(resource)
-            if not resource_.has_download():
-                collection_to_download_need_curation[collection_.identifier].append(resource)
-
-    # who is used more than once?
-    counter = Counter(prefix for c in nfdi_collections.values() for prefix in c.get_prefixes())
-
-    # first party
-    first_party = defaultdict(list)
-    for collection_ in nfdi_collections.values():
-        for prefix, call in manager.get_collection_first_party(
-            collection_, skip_org_rors={NFDI_ROR}
-        ).items():
-            if call:
-                first_party[prefix].append(collection_)
-    first_party_list = [
-        (manager.registry[prefix], consortia) for prefix, consortia in sorted(first_party.items())
-    ]
-
-    return render_template(
-        "nfdi.html",
-        collections=nfdi_collections,
-        tib_collection_mappings=tib_collection_mappings,
-        bartoc_collection_mappings=bartoc_collection_mappings,
-        collection_to_tib_opportunities=collection_to_tib_opportunities,
-        tib_opportunities=tib_opportunities,
-        prefix_counter=counter,
-        collection_to_license_needs_curation=collection_to_license_needs_curation,
-        collection_to_domain_needs_curation=collection_to_domain_needs_curation,
-        collection_to_download_need_curation=collection_to_download_need_curation,
-        first_party=first_party_list,
-        sort_collections=_sort_collections,
-    )
-
-
-def _sort_collections(c: list[Collection]) -> list[Collection]:
-    return sorted(c, key=_collections_key, reverse=True)
-
-
-def _collections_key(c: Collection) -> tuple[int, int]:
-    rv = len(c.maintainers or [])
-    rv += sum(
-        r.startswith("https://discord") or r.startswith("https://go.rocket")
-        for r in c.references or []
-    )
-    return rv, len(c.resources)
